Adapt AotProcessor to changes in Spring Framework
Closes gh-32560
