+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/java/org/springframework/boot/gradle/tasks/aot/ProcessAot.java
@@ -38,7 +38,7 @@ public class ProcessAot extends AbstractAot {
 
 	public ProcessAot() {
 		this.applicationClass = getProject().getObjects().property(String.class);
-		getMainClass().set("org.springframework.boot.AotProcessor");
+		getMainClass().set("org.springframework.boot.SpringApplicationAotProcessor");
 	}
 
 	@Input
